Core: Delete uninstall option for non xtecadmin user

+++ lib/classes/plugininfo/gradingform.php
@@ -31,7 +31,13 @@ defined('MOODLE_INTERNAL') || die();
 class gradingform extends base {
 
     public function is_uninstall_allowed() {
-        return true;
+        //XTEC ************ AFEGIT - Disable uninstalling
+        //2014.09.09  @pferre22
+        if (!get_protected_agora()) {
+            return false;
+        }
+        //************ FI
+        return false;
     }
 
     /**
@@ -39,22 +45,30 @@ class gradingform extends base {
      * This is intended for disabling of plugin, some DB table purging, etc.
      */
     public function uninstall_cleanup() {
-        global $DB;
+        //XTEC ************ AFEGIT - Disable uninstalling
+        //2018.05.28  @svallde2
+        if (!get_protected_agora()) {
+        //************ FI
+            global $DB;
 
-        // Find all definitions and templates.
-        $definitions = $DB->get_fieldset_select('grading_definitions', 'id', 'method = ?', [$this->name]);
-        if ($definitions) {
-            // Delete instances and definitions. Deleting instance will not delete grades because they were
-            // already pushed to the module and gradebook.
-            list($sqld, $paramsd) = $DB->get_in_or_equal($definitions);
-            $DB->delete_records_select('grading_instances', 'definitionid ' . $sqld, $paramsd);
-            $DB->delete_records_select('grading_definitions', 'id ' . $sqld, $paramsd);
+            // Find all definitions and templates.
+            $definitions = $DB->get_fieldset_select('grading_definitions', 'id', 'method = ?', [$this->name]);
+            if ($definitions) {
+                // Delete instances and definitions. Deleting instance will not delete grades because they were
+                // already pushed to the module and gradebook.
+                list($sqld, $paramsd) = $DB->get_in_or_equal($definitions);
+                $DB->delete_records_select('grading_instances', 'definitionid ' . $sqld, $paramsd);
+                $DB->delete_records_select('grading_definitions', 'id ' . $sqld, $paramsd);
+            }
+            // Delete templates for this grading method.
+            $DB->delete_records_select('grading_areas', 'component = ? AND activemethod = ?', array('core_grading', $this->name));
+            // Update the remaining grading areas to use simple grading method instead of this grading method.
+            $DB->execute('UPDATE {grading_areas} SET activemethod = NULL WHERE activemethod = ?', [$this->name]);
+
+            parent::uninstall_cleanup();
+        //XTEC ************ AFEGIT - Disable uninstalling
+        //2018.05.28  @svallde2
         }
-        // Delete templates for this grading method.
-        $DB->delete_records_select('grading_areas', 'component = ? AND activemethod = ?', array('core_grading', $this->name));
-        // Update the remaining grading areas to use simple grading method instead of this grading method.
-        $DB->execute('UPDATE {grading_areas} SET activemethod = NULL WHERE activemethod = ?', [$this->name]);
-
-        parent::uninstall_cleanup();
+        //************ FI
     }
 }
